requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
pyarr>=5.2.0
